refactor(nav-bar): Renamed task bar directory to nav bar
