+++ src/app/components/nav-bar/nav-bar.scss
